+++ lib/zxcvbn/score.rb
@@ -2,12 +2,11 @@
 
 module Zxcvbn
   class Score
-    attr_accessor :guesses, :entropy, :crack_time, :crack_time_display, :score,
+    attr_accessor :guesses, :crack_time, :crack_time_display, :score,
                   :match_sequence, :password, :calc_time, :feedback
 
     def initialize(options = {})
       @guesses            = options[:guesses]
-      @entropy            = options[:entropy]
       @crack_time         = options[:crack_time]
       @crack_time_display = options[:crack_time_display]
       @score              = options[:score]